requirements.txt
flask==3.0.0
pdfplumber==0.10.3
openai>=1.50.0
tavily-python==0.5.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
gunicorn==21.2.0
